SSH_Access: be more liberal about recognizing password prompts
A SUSE customer reported an issue where the use of a 3rd party
PAM module in stack changed the password prompt to:

    UNIXADS password for username:

Unfortunately, the ssh parsing code in lftp does not recognize
this as a password prompt, which means that the login process
gets stuck.

We have found that accepting all strings that contain the phrase
"password" and end in a colon works very well and it also covers
this particular case.