Ensure that PollVec enum matches poll values
IN and OUT are used by PollVec mask functions to determine which fd_sets
to use in AddFD, FDReady, and FDSetNotReady, however these enums are not
used outside of PollVec code and the rest of the code uses POLLIN and
POLLOUT definitions instead. This works on many POSIX platforms since
the enum values match the POLLIN and POLLOUT values, but not on AIX
where POLLOUT is 2 instead of 4. This prevents code from determining
when a file descriptor is writeable, leading to a hang when using FTPS
protocol (at least).

Fixes #672