fix stack corruption in `cd` for relative paths due to added prefix length
when the current path is not absolute, the implementation prefixes
the target directory to change to with ~/
this is 2 more characters than allocated, hence we write beyond
the buffer boundary.

in order to fix this, we allocate two more characters for the reconstruction
buffer, so it can hold a relative path prefixed with ~/ as well.